[refactor] Create Chat dataclass to hold all of chat state
We had three variables for the current chat in the session_state
and it was beginning to get a bit messy. So as I wanted
to change the "created_at" variable saved to disk to be set to when
New Chat was pressed, adding more state, I decided it was time to
create a class.