[refactor] Create strongly typed state in view_chat
We are getting a lot of stuff in the session state, and it's not
described anywhere. We can get the type system to represent this
by creating a class for it, State, and then telling the type system
that we want to treat the session state as that class via:

_s = cast(State, st.session_state)

As session_state magics its contents into attributes, when we call
the State attributes they are really just calls into st.session_state
and so work fine. It's a bit easier than creating a wrapper class. But
not 100% sure it's not a bit evil.