[feat] remember last used model across restarts
Fixes: https://github.com/mikerhodes/interlocution/issues/10

Add a configuration file to `~/.interlocution/config.json`.
When the model is changed, save it to `last_used_model`.
On starting, load up the config and use the last_used_model if it's
not none and it's in our list. Otherwise revert to first model in the
list.

Right now we read the file every time we use it. This doesn't seem
awful right now, albeit it's pretty gratuitous.