[fix] Loading chat from history page doesn't save model to config
Create a "protocol" between the chat history page and the chat view
page so the chat history page can instruct the chat view page to
load a chat. This allows us to use the same code to load a chat from
both places.

This fixes a bug where loading from the history page didn't save the
model from the chat as the last used model because I forgot to update
the load_chat code to save the model in the history page. It made sense
to do the larger refactor.