[feat] Use types for chat messages
It was becoming hard to remember the structure of each type of message.
Sadly, this will break old saved chats, but I don't think anyone is
using this apart from me.

I also want to hang more information from messages, such as the model
that generated them and the created date.