[feat] Auto-update chats exported to obsidian
When a chat has been saved to obsidian, carrying on the chat will now
update the copy in obsidian.

Add an attribute to Chat, export_location. Set this when a chat is
saved to Obsidian. When the Chat is saved, also write a markdown copy
to this location.

Save and load the new attribute to/from JSON.