[refactor] manage context length in chat gateway adaptors
Managing the context length in the view_chat module didn't make that
much sense, and wasn't used at all in the Anthropic or Watsonx
adapters --- because its main use is to control local model memory
use. So just stick it into the Ollama adaptor.

This also means we can get rid of show from the ChatAdaptor protocol,
as it's not used anywhere but internally for the Ollama adaptor.