[refactor] Convert history store to pydantic: human fixes
- update uv.lock
- Remove BaseMessage base class as not correct for IncludedFile.
- Remove unneeded type conversion (eg to datetime) before deserialising
  chats (pydantic will handle str to datetime).
- Add discriminator to *Message types by adding the type field to the
  model classes and telling pydantic to use that to figure out the type
  to deserialise into from the `messages` MessageList union. This
  removes the code in chathistory.py that injected the type field and
  then used it to deserialise (AI left that step in rather than
  leveraging pydantic).

Overall the AI did a good job in starting the conversion, but did not
use several useful, more advanced pydantic features that could remove
code from rapport.

I'm still not 100% convinced that going to pydantic here is the right
choice, but it should allow me to add and remove fields from the Chat
and Message classes more easily, which is good as I'd been a bit
unwilling to do so because of the annoyance of updating the
serialisation code.