[refactor] move source files to src/rapport
It turns out that the python files for the package should be in
src/package_name. Otherwise the entrypoint won't work because the
python files in src don't get packaged into the rapport package.

Overall, this feels like something I should've known, and perhaps did
know once, but forgot what with writing predominantly Go for five years.

I have half an idea to publish the rapport package so it can be used
via things like `uv tool` so hopefully this brings us closer to that.