fix: delete images using chat id prefix
Images for a chat are saved using their chat id as a prefix. So we can
delete the images for a chat by globbing for the files with the
chat id prefix rather than loading the chat and looking for included
images.

The AI couldn't do it that way because it didn't include the chat id
in the file name, but I discovered that led to a bug where if two chats
included the same image, if you deleted one the other would lose its
image. So I added the ID to the image file names, which helped simplify
this code too.