fix: display smaller thumbnails of images in chat
By adding columns into the chat message for included images, we can
make the included images take up much less vertical space. Usefully,
this is better than the `width` option to st.image, because the
original image is used when you click the widget to show the whole
image instead of the resized. Ace!