feat: support openai latex markup for equations
A nice thing that recent openai models will do is return latex
formatted equations. This is very helpful when asking about details
of neural networks! And obviously in other places but this is where
I've been seeing them the most.

Streamlit supports latex, but with different start/end markers.

So we do a couple of things:

- Sometimes models will return the two-character models as multiple
  chunks. So take the opportunity to collect larger chunks of text
  before yielding them to streamlit both to ensure we get the full
  start/end marker and also because after seeing the granularity
  of the chunks (single tokens mostly) I realised it'd be good to
  rerender the UX for larger chunk values. 60 felt like a good chunk
  size in use, so I used that.
- Second, postprocess each chunk to switchout the openai style markers
  for streamlit markdown compatible ones.

I chose not to add the processing for older chats in the end as it
just didn't feel quite right or necessary to repeatedly postprocess
those chats, particularly as most of them don't need it at all.