feat: request latex from chat models (like Claude)
Add an "extras" set of instructions to allow us to add to the
system prompt created by the user. This means that systemprompt.md
can be more "what the user wants" than a set of instructions that
help for Rapport's rendering.

Add a first instruction to return equations as latex, which appears
to also enable latex inclusion in Claude responses. It'd be nice if
there was an API switch for this, but for now adding to the system
prompt seems to work nicely. And Claude returns using streamlit
compatable $$ markers, meaning that there are no rendering changes
needed.

Neat.