feat: make generic system prompt and allow customisation
Instead of allowing the user to replace the system prompt, instead allow
customisation via extra content to append to it. This means that the
basic personality can stay but users can customise it how they like.

Expand the system prompt using Anthropic's as an inspiration.

Add the current date.

Add a config setting for extra information to add to the system prompt.