fix: don't stop rendering messages on streamlit reload
Fixes: https://github.com/mikerhodes/rapport/issues/27

When changing code, often messages will stop rendering in the chat
history and will not be processed correctly when creating messages
to send to the AI model. I think this is because somehow the classes
loaded by the chat history become "out of sync" with the types used
in the match statement.

Specifically, I found that directly matching on the type using `case
SystemMessage()` failed after streamlit reloads. I'm not exactly sure
why, but it meant that both chat history rendering failed, and preparing
messages to send to the model.

This made updating Rapport more of a chore because you'd have to
restart the app (usually after a moment of staring at a blank screen)
in order to test new changes. Not the end of the world but repeatedly
frustrating, which isn't great.

I also tried out using isinstance(o, SystemMessage) but that had the
same problem.