feat: support tool-calling via MCP
This commit adds support for calling tools provided by MCP servers.
The servers can either be reached over stdio, in which case Rapport
starts a child process and communicates over stdio, or HTTP, in which
case Rapport expects a server to be already running and exposing a
HTTP endpoint. The commit implements an allowlist approach for safety,
allowing only the tools in that list to be used by a model.

First, a user configures MCP servers by writing JSON into the config
screen set up by view_config.py. This is a bit fiddly but it'll do.
The pydantic errors are inscruitable but we don't leave the JSON
invalid. There is inline help, and an expander that will show active
tools for debugging.

Second, tools.py contains the code to create a ToolRegistry. This is
initialised with the configuration for MCP servers. It queries each
server for its available tools and compares them with the allowlist
configured for the server. Those tools on the list are added to the
registry.

The registry also provides for calling the tools found during
initialisation.

Tool calls are performed via the fastmcp library's client. This client
is async, so we do a basic wrapping of calls that use it via
asyncio.run(). This is a bit inefficient as, due to the library
design, we have to create a new connection each time. This isn't too bad
for the HTTP servers, but for stdio it means starting the server
each time we call a tool. Not great, but we can fix that later if
it's a problem.

Finally, we alter view_chat so that tool calls can be shown inline
with assistant chat messages, in a single chat block. This involves
creating tool request and response message types in chat model.
Rendering a chat is altered such that as we read the chat we are able
to place subsequent assistant or user messages in the same chat block.
We do this by creating a peekable iterator using more_itertools
package, which allows us to write methods that will consume all the
messages in each block while rendering them, and then when we peek
the iterator and find that it's flipped participant, we can return
and have the next item handled in the parent method.

Regenerating the assistant message doesn't quite work properly yet. It
doesn't reverse over all the assistants messages (which can now
include tool calls and responses). Overall things work okay though,
and we can fix it up later.

Finally we add two demo MCP servers, and use them in the documentation
for adding a stdio and a HTTP server. These are mostly for debugging,
although they provide mildly useful tools.