tools: load clients sync during initialise
We only call ToolRegistry.initialise once, at startup. fastmcp Clients
don't need to be created in an aysnc context, only used in one.
We can simplify things by creating all the clients during initialise,
meaning we have a less nested get client method and can get rid of
the lock we used to protect the client cache, as it's can never be
read concurrently to being written due to the code flow.